Use postgres-interval to unmap intervals
This is IMO superior - the code is shorter, (mostly) clearer, and it
leverages the lib for the corner cases. In particular, it handles
floats at every position without generating a malformed postgres
interval string, which the previous version didn't.

With this change, it's feasible to make all of the fields in the input
into `Floats`, which I think would be backwards compatible. It still
makes sense for the `Interval` type itself to have all `Int` fields
except for milliseconds.