Allow more specific `TArgs` types
Allow types more specific than `{ [argName: string]: any }` to be
passed into `makeWrapResolversPlugin`. We now set `TSource`, `TContext`
and `TArgs` as type parameters to `makeWrapResolversPlugin` which
means we can optionally pass in types more narrow than the defaults.

This _should_ be a backwards compatible change since we set the type
defaults to be the same as in `ResolverWrapperFn`. Fixes #708.